feat:login관련 dto,excption 추가 및 user controller, service 수정

+++ src/main/java/kbits/kb04/parasol/users/repository/UserAssetRepository.java
@@ -20,7 +20,7 @@ import kbits.kb04.parasol.users.entity.Users;
 이러한 메서드를 사용하여 데이터베이스와 상호작용할 수 있으며, 별도의 SQL 쿼리를 작성할 필요가 없습니다. 
 Repository 인터페이스를 사용하여 데이터베이스 연동 코드를 간단하게 작성할 수 있습니다.*/
 
-public interface UserAssetRepository extends JpaRepository<UserAsset, Integer> {
-	Optional<UserAsset> findByUserNo(int no);
+public interface UserAssetRepository extends JpaRepository<UserAsset, Long> {
+	Optional<UserAsset> findByUserNo(Long no);
     
 }
